requirements.txt
# Creative Research Report — LLM + MCP + Scraping + Storage
# Core
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# Scraping
apify-client>=1.6.0
google-api-python-client>=2.100.0
tavily-python>=0.3.0
# Reddit: public JSON API only (no praw)

# Storage
pyairtable>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0

# MCP server (optional)
mcp[cli]>=1.0.0
